Enables MN68820 rangefinder on SpeedyBeeF405WING
Adds support for the MN68820 rangefinder to the SpeedyBeeF405WING board.

This includes enabling the rangefinder in the hwdef.dat, setting the
default I2C address, and implementing probe, init, and read functions
based on the STM32F103 implementation, complete with detailed debug
messages. This also includes the firmware upload logic.